First step of  moving RichLanguage into scafi.lib
