+++ core/src/main/scala/it/unibo/scafi/lib/DynamicCode.scala
@@ -42,8 +42,6 @@ trait StdLib_DynamicCode {
       foldhood(injecter())((f1,f2) => if(f1.ver>=f2.ver) f1 else f2)(nbr{f})
     }
 
-    import Builtins.Bounded
-
     def exec[T,R:Bounded](procs: List[Fun[T,R]], arg: T, maxVer: Int, curVer: Int, numNbrs: Int, Null: R): (R,Int) = {
       val Fun(headVer,headFun) = procs.head
       // If the version of the head of the list is >= of the min version exec by any neighbour, let's run it
