+++ demos/src/main/scala/sims/DISIDemo.scala
@@ -18,8 +18,9 @@
 
 package sims
 
-import it.unibo.scafi.incarnations.BasicSimulationIncarnation.{ AggregateProgram, Builtins }
+import it.unibo.scafi.incarnations.BasicSimulationIncarnation.{ AggregateProgram }
 import it.unibo.scafi.simulation.gui.{Launcher, Settings}
+import it.unibo.scafi.lib.Bounded
 
 object DISIDemo extends Launcher {
   Settings.Sim_ProgramClass = "sims.Main" + (if(args.length == 0) "" else args(0))
@@ -87,7 +88,7 @@ class Main11 extends DISIDemoAggregateProgram {
 }
 
 class Main12 extends DISIDemoAggregateProgram {
-  import Builtins.Bounded.of_i
+  import Bounded.of_i
 
   override def main() = maxHoodPlus(boolToInt(nbr{sense1}))
 }
@@ -97,7 +98,7 @@ class Main13 extends DISIDemoAggregateProgram {
 }
 
 class Main14 extends DISIDemoAggregateProgram {
-  import Builtins.Bounded.of_i
+  import Bounded.of_i
 
   override def main() = rep(0){ x => boolToInt(sense1) max maxHoodPlus( nbr{x}) }
 }
