+++ demos/src/main/scala/sims/SpawnTest.scala
@@ -19,6 +19,7 @@
 package sims
 
 import it.unibo.scafi.incarnations.BasicSimulationIncarnation._
+import it.unibo.scafi.lib.Bounded
 import it.unibo.scafi.simulation.gui.{Launcher, Settings}
 
 object SpawnTestRunner extends Launcher {
@@ -80,8 +81,6 @@ class SpawnTest extends AggregateProgram with SensorDefinitions with FieldUtils 
       }.value
     }
   }
-
-  import Builtins.Bounded
   def minHoodSelector[T: Bounded, V](toMinimize: => T)(data: => V): Option[V] = {
     val ord = implicitly[Bounded[T]]
     foldhoodPlus[(T,Option[V])]((ord.top, None))( (x,y) => if(ord.compare(x._1,y._1) <= 0) x else y )((toMinimize, Some(data)))._2
