+++ simulator/src/main/scala/it/unibo/scafi/incarnations/SimulationIncarnations.scala
@@ -18,7 +18,7 @@
 
 package it.unibo.scafi.incarnations
 
-import it.unibo.scafi.lib.StandardLibrary
+import it.unibo.scafi.lib.{Bounded, StandardLibrary}
 import it.unibo.scafi.simulation.{Simulation, SpatialSimulation}
 import it.unibo.scafi.space.BasicSpatialAbstraction
 import it.unibo.scafi.time.BasicTimeAbstraction
@@ -35,8 +35,7 @@ object BasicSimulationIncarnation
   extends BasicAbstractSimulationIncarnation
     with StandardLibrary {
 
-  import Builtins.Bounded
-  override implicit val idBounded: Bounded[ID] = Builtins.Bounded.of_i
+  override implicit val idBounded: Bounded[ID] = Bounded.of_i
 }
 
 class BasicAbstractSpatialSimulationIncarnation
