+++ tests/src/test/scala/it/unibo/scafi/test/FunctionalTestIncarnation.scala
@@ -19,10 +19,9 @@
 package it.unibo.scafi.test
 
 import it.unibo.scafi.incarnations.AbstractTestIncarnation
-import it.unibo.scafi.lib.StandardLibrary
+import it.unibo.scafi.lib.{Bounded, StandardLibrary}
 import it.unibo.scafi.simulation.Simulation
 
 object FunctionalTestIncarnation extends AbstractTestIncarnation with Simulation with StandardLibrary {
-  import Builtins.Bounded
-  override implicit val idBounded: Bounded[ID] = Builtins.Bounded.of_i
+  override implicit val idBounded: Bounded[ID] = Bounded.of_i
 }
